# docs/openapi/http-utilities.yaml
openapi: 3.0.3
info:
  title: LerianStudio lib-commons HTTP Utilities
  description: |
    OpenAPI specification for HTTP utility functions provided by the LerianStudio lib-commons library.
    
    This specification documents the standardized HTTP response patterns and middleware components
    available for building consistent APIs using the lib-commons library.
    
    ## Key Features
    - **Standardized Response Functions**: Consistent HTTP status code handlers
    - **Error Response Format**: Unified error structure across all services
    - **Middleware Components**: CORS, authentication, logging, telemetry
    - **Health Check Framework**: Standardized health monitoring endpoints
    - **Pagination Support**: Cursor-based pagination helpers
    
    ## Usage Pattern
    ```go
    import httpCommons "github.com/LerianStudio/lib-commons/commons/net/http"
    
    // In your Fiber handlers
    func GetUser(c *fiber.Ctx) error {
        user, err := userService.GetUser(c.Params("id"))
        if err != nil {
            return httpCommons.NotFound(c, "USER_001", "User Not Found", err.Error())
        }
        return httpCommons.OK(c, user)
    }
    ```
  version: 1.0.0
  contact:
    name: LerianStudio
    url: https://github.com/LerianStudio/lib-commons
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.example.com/v1
    description: Example API server using lib-commons
  - url: http://localhost:8080
...
paths:
...
  /users/{id}:
    get:
      tags:
        - response-patterns
      summary: Example User Retrieval (Success Pattern)
      description: |
        Example demonstrating successful response pattern using `httpCommons.OK()`.
        
        Implementation:
        ```go
        func GetUser(c *fiber.Ctx) error {
            user, err := userService.GetUser(c.Params("id"))
            if err != nil {
                return httpCommons.NotFound(c, "USER_001", "User Not Found", err.Error())
            }
            return httpCommons.OK(c, user)
        }
        ```
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              examples:
                successful_user:
                  summary: Successfully retrieved user
                  value:
                    id: "550e8400-e29b-41d4-a716-446655440000"
                    name: "John Doe"
                    email: "john.doe@example.com"
                    created_at: "2024-01-15T10:00:00Z"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                user_not_found:
                  summary: User not found error
                  value:
                    code: "USER_001"
                    title: "User Not Found"
                    message: "User with ID 550e8400-e29b-41d4-a716-446655440000 does not exist"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                internal_error:
                  summary: Database connection error
                  value:
                    code: "SYS_001"
                    title: "Internal Server Error"
                    message: "Database connection failed"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      description: |
        Standard error response format used by all lib-commons HTTP response functions.
        
        Maps to `commons.Response` struct with consistent field naming.
      properties:
        code:
          type: string
          description: Error code for programmatic handling
          example: "USER_001"
        title:
          type: string
          description: Human-readable error title
          example: "User Not Found"
        message:
          type: string
          description: Detailed error message
          example: "User with ID 550e8400-e29b-41d4-a716-446655440000 does not exist"
        entityType:
          type: string
          description: Optional entity type that caused the error
          example: "User"
      required:
        - code
        - title
        - message
...
    User:
      type: object
      description: User entity example
      properties:
        id:
          type: string
          format: uuid
          description: User unique identifier
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          description: User full name
          example: "John Doe"
        email:
          type: string
          format: email
          description: User email address
          example: "john.doe@example.com"
        created_at:
          type: string
          format: date-time
          description: User creation timestamp
          example: "2024-01-15T10:00:00Z"
      required:
        - id
        - name
        - email
        - created_at
...